Extract folder rename tree updates
